+++ src/components/extension/ShareIntentRouter.tsx
@@ -1,13 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import { BackHandler, Linking } from 'react-native';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import DoLinkWebView from '../DoLinkWebView';
 import type { ShareIntentData } from '../../types/shareIntent';
+import DoLinkShareIntentView from '../DoLinkShareIntentView';
 
 /**
  * Share Intent 진입 시 웹뷰를 띄워 로그인 여부를 브릿지로 확인하는 라우터
  */
 export default function ShareIntentRouter(props: ShareIntentData) {
+  console.log('ShareIntentRouter', props);
   const [shareIntent, setShareIntent] = useState<ShareIntentData | null>(null);
 
   // 로그인 여부는 DoLinkWebView 내부의 브릿지 통신을 통해 확인합니다.
@@ -26,5 +25,5 @@ export default function ShareIntentRouter(props: ShareIntentData) {
 
   // DoLinkWebView를 항상 렌더링하며, 내부에서 로그인 여부에 따라 분기 처리함
 
-  return <DoLinkWebView shareIntent={shareIntent} />;
+  return <DoLinkShareIntentView shareIntent={shareIntent} />;
 }
